requirements.txt
nest-asyncio
torch
fastapi
python-dotenv
llama-index
llama-parse
llama-index-vector-stores-chroma
chromadb
starlette